# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pesisir Project",
        "description": "API Documentation for Pesisir",
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get list of activities",
                "operationId": "53e7a0b67f6097ed90245000764b2e5e",
                "responses": {
                    "200": {
                        "description": "List of activities with locations and volunteers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Create a new activity",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date",
                                    "time",
                                    "status",
                                    "fee",
                                    "loc_name",
                                    "loc_address",
                                    "lat",
                                    "long"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Beach Cleanup"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Cleaning the beach area"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-01"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ongoing",
                                            "done",
                                            "upcoming"
                                        ],
                                        "example": "upcoming"
                                    },
                                    "fee": {
                                        "type": "integer",
                                        "example": 20000
                                    },
                                    "loc_name": {
                                        "type": "string",
                                        "example": "Pantai Parangtritis"
                                    },
                                    "loc_address": {
                                        "type": "string",
                                        "example": "Jl. Pantai Selatan, Bantul"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -7.9778
                                    },
                                    "long": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 110.3695
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Activity created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get a single activity",
                "operationId": "6d75cdd5535b9b12358448ec198faba4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity details"
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Update an activity",
                "operationId": "064379988a5fe2c974ae5ab18f0cddfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Activity"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-10"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ongoing",
                                            "done",
                                            "upcoming"
                                        ],
                                        "example": "ongoing"
                                    },
                                    "fee": {
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "loc_name": {
                                        "type": "string",
                                        "example": "New Location"
                                    },
                                    "loc_address": {
                                        "type": "string",
                                        "example": "Jl. Malioboro, Yogyakarta"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -7.8014
                                    },
                                    "long": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 110.3647
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Activity updated"
                    },
                    "404": {
                        "description": "Activity not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Delete an activity",
                "operationId": "d822403ad9a85feb951b1073858a072d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity deleted"
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}